Module unsafe_tips_and_tricks::closures

Functions

execute_a_closure⚠
An arbitrary function which executes the provided function, using a pointer to some user_data for context.

raw_callback
Get a function pointer which can be used as a Callback that accepts a pointer to the closure as its user_data.

Type Definitions

Callback