Function unsafe_tips_and_tricks::closures::raw_callback

pub fn raw_callback<F>(_closure: &F) -> Callback where
    F: FnMut(i32) -> i32, 

Get a function pointer which can be used as a Callback that accepts a pointer to the closure as its user_data.